Split DataService into two smaller services
- StateManagerService responsible for handling board/window operations
- DatabaseService responsible for communicating with Firebase's database
