+++ src/app/components/window-dialog/window-dialog.component.ts
@@ -1,6 +1,6 @@
 import { Component, Inject } from '@angular/core';
 import {MatDialog, MatDialogRef, MAT_DIALOG_DATA} from '@angular/material';
-import {DataService} from '../../providers/data.service';
+import { StateManagerService } from '../../providers/state-manager.service';
 
 @Component({
   selector: 'app-window-dialog',
@@ -10,7 +10,7 @@ import {DataService} from '../../providers/data.service';
 export class WindowDialogComponent {
   private _selectedFile: string;
 
-  constructor(private _dataService: DataService, 
+  constructor(private stateManagerService: StateManagerService, 
               public dialogRef: MatDialogRef<WindowDialogComponent>, 
               @Inject(MAT_DIALOG_DATA) public data: any) {
     this._selectedFile = "";
@@ -22,8 +22,7 @@ export class WindowDialogComponent {
   }
 
   createWindow(name): void {
-    this._dataService.addWindow(name, this._selectedFile);
-    this._dataService.getBoards();
+    this.stateManagerService.addWindow(name, this._selectedFile);
     this.dialogRef.close();
   }
