+++ src/app/modules/shared/shared.module.ts
@@ -3,6 +3,7 @@ import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
 import { ModuleWithProviders } from '@angular/core';
 import { AuthenticationService } from '../../modules/login/providers/authentication.service';
+import { StateManagerService } from '../../providers/state-manager.service';
 import { BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import { MatButtonModule, MatCheckboxModule, MatDialogModule, MatToolbarModule, 
         MatTabsModule, MatButtonToggleModule, MatIconModule, MatSidenavModule,
@@ -41,7 +42,8 @@ export class SharedModule {
     return {
       ngModule: SharedModule,
       providers: [
-        AuthenticationService
+        AuthenticationService,
+        StateManagerService
       ]
     };
   }
